v1.2.3-01: split app panel
