+++ src/jsMain/kotlin/top/hikki/hikkitop/Fragment/ReverserPanel.kt
@@ -1,0 +1,64 @@
+package top.hikki.hikkitop.Fragment
+
+import io.kvision.form.text.textAreaInput
+import io.kvision.html.button
+import io.kvision.html.div
+import io.kvision.html.p
+import io.kvision.panel.SimplePanel
+import kotlinx.browser.document
+import kotlinx.coroutines.launch
+import org.w3c.dom.HTMLElement
+import org.w3c.dom.HTMLTextAreaElement
+import top.hikki.hikkitop.AppScope
+import top.hikki.hikkitop.Model.ReverserModel
+
+internal val reverserPanel = SimplePanel {
+    p(){setAttribute("id", "textAreaInfo")}
+    div(className = "form-group") {
+        val textInputInfo = document.getElementById("textAreaInfo")
+        textInputInfo?.innerHTML = "<p>Input Text <font size=2>(Based on <a href='https://cloud.baidu.com/product/speech/tts_online'>Baidu TTS</a>)</font></p>"
+        textAreaInput(rows=3, className = "form-control") {
+            setAttribute("id", "textArea1")
+            setAttribute("maxlength", "50")
+            setAttribute("placeholder", "哎呀米诺")
+        }
+    }
+    div(className = "btn-set") {
+        button("Convert", className = "btn btn-secondary long-btn") {
+            onClick {
+                val textAreaE = document.getElementById("textArea1") as HTMLTextAreaElement
+                AppScope.launch {
+                    val barElement = document.getElementById("process-bar") as HTMLElement
+                    barElement.hidden = false
+                    barElement.innerHTML = "<p class='text-warning' style='float:left;'><b>processing...</b></p><img src='static/Pulse.gif' width='50px'>"
+
+                    val audioBase64 = ReverserModel.getAudioByText(textAreaE.value)
+                    val audioContainer = document.getElementById("audio-container") as HTMLElement
+                    audioContainer.innerHTML = "<audio id='audio-player' controls><source type='audio/wav' src='$audioBase64'></audio>"
+                    barElement.innerHTML = "<img src='static/done.png' width='40px' style='float:left'><p class='text-success'><b>Done</b></p>"
+                }
+            }
+        }
+    }
+}
+
+internal val recordPanel = SimplePanel {
+    div(className = "btn-set") {
+        button("Convert", className = "btn btn-secondary long-btn") {
+            onClick {
+                AppScope.launch {
+                    val barElement = document.getElementById("process-bar") as HTMLElement
+                    barElement.hidden = false
+                    barElement.innerHTML = "<p class='text-warning' style='float:left;'><b>processing...</b></p><img src='static/Pulse.gif' width='50px'>"
+
+                    val base64 = document.getElementById("recordingsList") as HTMLElement
+                    val base64Str = base64.innerText.split(",")[1]
+                    val audioBase64 = ReverserModel.getAudioByFile(base64Str)
+                    val audioContainer = document.getElementById("audio-container") as HTMLElement
+                    audioContainer.innerHTML = "<audio id='audio-player' controls><source type='audio/wav' src='$audioBase64'></audio>"
+                    barElement.innerHTML = "<img src='static/done.png' width='40px' style='float:left'><p class='text-success'><b>Done</b></p>"
+                }
+            }
+        }
+    }
+}
